Updates include (fix UI bug & optimization): (DataTable) - create utils/Table to store all table parts instead of including them directly in DataTable file - include the Border component (previously SortBorder) and fix the icon alignment in utils/SortBy - increase width and z-index of utils/Resizer to prevent conflicts with utils/SortBy on resize - resolve the table’s scroll bar (horizontal/vertical) behavior - optimize JSX in the DataTable component - use shared/FilterTextInput in utils/GlobalFiter (Other) - create shared/FilterTextInput - fix the expanded table view z-index issue - use shared/FilterTextInput in SamplesTable/cells/AdditionalMetadata modal - add position:relative to shared/BoxBlock - make minor adjustment in SamplesTable based on the updates
