+++ src/components/shared/BoxBlock.js
@@ -2,7 +2,7 @@ import { Box } from 'grommet'
 
 export const BoxBlock = ({ children, style, ...props }) => (
   // eslint-disable-next-line react/jsx-props-no-spreading
-  <Box style={{ display: 'block', ...style }} {...props}>
+  <Box style={{ display: 'block', position: 'relative', ...style }} {...props}>
     {children}
   </Box>
 )
