+++ src/components/shared/DataTable/DataTable.js
@@ -1,3 +1,5 @@
+/* eslint-disable no-nested-ternary */
+/* eslint-disable react/jsx-props-no-spreading */
 import { useEffect, useState, useRef, memo } from 'react'
 import {
   useFlexLayout,
@@ -8,90 +10,17 @@ import {
 import { useSticky } from 'react-table-sticky'
 import { useIntersectObserver } from 'hooks/useIntersectObserver'
 import { Box, Text } from 'grommet'
-import styled, { css } from 'styled-components'
+import { BoxBlock } from 'components/shared/BoxBlock'
 import {
   HorizontalScrollIndicator,
   Resizer,
   SortBy,
-  SortByBorder
+  Table,
+  TableHeader,
+  TableBody,
+  TableRow,
+  TableCell
 } from './utils'
-
-const borderColor = 'gray-shade-20'
-const alternateRowBg = 'gray-shade-5'
-const rowHoverBG = '#E2E2E2'
-const headerHeight = '44px'
-
-const Table = styled(Box)`
-  overflow: scroll;
-  scrollbar-width: none;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-  [data-sticky-td] {
-    position: sticky;
-  }
-  ${({ theme }) => css`
-    [data-sticky-last-left-td] {
-      box-shadow: 5px 2px 15px ${theme.global.colors[borderColor]};
-    }
-    [data-sticky-first-right-td] {
-      box-shadow: -2px 0px 3px ${theme.global.colors[borderColor]};
-    }
-  `}
-`
-
-const TableHeader = styled(Box)`
-  position: sticky;
-  top: 0;
-  z-index: 1;
-`
-
-const TableBody = styled(Box)`
-  position: relative;
-  z-index: 0;
-`
-
-const TableRow = styled(Box)`
-  ${({ theme }) => css`
-    &:last-child div {
-      border-bottom: 0;
-    }
-    &:nth-child(even) {
-      > div {
-        background: ${theme.global.colors[alternateRowBg]};
-      }
-      &:hover div {
-        background: ${rowHoverBG};
-      }
-    }
-  `}
-`
-
-const TableCell = styled(Box)`
-  ${({ theme }) => css`
-    background: ${theme.global.colors.white};
-    border-bottom: 1px solid ${theme.global.colors[borderColor]};
-    border-right: 1px solid ${theme.global.colors[borderColor]};
-    overflow: hidden;
-    padding: 8px 16px;
-    position: relative;
-    overflow: hidden;
-    &:last-child {
-      border-right: 0;
-    }
-    span {
-      font-size: 14px;
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
-  `}
-
-  ${({ theme, type }) =>
-    type === 'th' &&
-    css`
-      box-shadow: 5px 2px 15px ${theme.global.colors[borderColor]};
-    `}
-`
 
 export const DataTable = ({
   columns,
@@ -100,8 +29,7 @@ export const DataTable = ({
   hiddenColumns = [],
   manualPagination = false,
   tableExpanded,
-  tableDefaultHeight, // required for the TableBody y-scroll
-  tableExpandedHeight
+  tableHeight // required for the expanded table view
 }) => {
   const tableRef = useRef(null)
   const firstCellRef = useRef(null)
@@ -114,7 +42,7 @@ export const DataTable = ({
   const isLastCellVisible = useIntersectObserver(lastCellRef, {
     root: tableRef.current,
     rootMargin: '0px',
-    threshold: 0.99
+    threshold: 0.8
   }).isIntersecting
 
   const tableInstance = useTable(
@@ -140,7 +68,6 @@ export const DataTable = ({
       columnResizing: { columnWidths }
     }
   } = tableInstance
-
   const [cellWidths, setCellWidths] = useState({})
 
   useEffect(() => {
@@ -155,102 +82,93 @@ export const DataTable = ({
         color: 'gray-shade-20',
         side: 'all'
       }}
+      height={{ min: '160px' }}
       margin={{
         horizontal: tableExpanded ? 'basex6' : 'none'
       }}
       style={{ overflow: 'visible' }}
     >
-      <Box style={{ position: 'relative' }}>
+      <BoxBlock>
         <HorizontalScrollIndicator
           isFirstCellVisible={isFirstCellVisible}
           isLastCellVisible={isLastCellVisible}
           target={tableRef.current}
         />
-        <Table
-          // eslint-disable-next-line react/jsx-props-no-spreading
-          {...getTableProps()}
-          ref={tableRef}
-          style={{ minWidth: 'auto' }}
+        <BoxBlock
+          width={{ max: 'none' }}
+          style={{
+            left: '-15px',
+            overflow: 'hidden',
+            width: 'calc(100% + 15px)'
+          }}
         >
-          <TableHeader className="header" style={{ width: '100%' }}>
-            {headerGroups.map((headerGroup) => (
-              <TableRow
-                // eslint-disable-next-line react/jsx-props-no-spreading
-                {...headerGroup.getHeaderGroupProps()}
-                direction="row"
-              >
-                {headerGroup.headers.map((column, i, arr) => (
-                  <TableCell
-                    // eslint-disable-next-line react/jsx-props-no-spreading
-                    {...column.getHeaderProps(column.getSortByToggleProps())}
-                    type="th"
-                    height={headerHeight}
-                    ref={
-                      // eslint-disable-next-line no-nested-ternary
-                      i === 0
-                        ? firstCellRef
-                        : i === arr.length - 1
-                        ? lastCellRef
-                        : null
-                    }
-                  >
-                    <Box direction="row" justify="between">
-                      <Text weight="bold">{column.render('Header')}</Text>
-                      {column.canSort && (
-                        <SortBy
-                          isSorted={column.isSorted}
-                          isSortedDesc={column.isSortedDesc}
+          <Table
+            {...getTableProps()}
+            ref={tableRef}
+            height={{ max: tableHeight }}
+          >
+            <TableHeader className="header">
+              {headerGroups.map((headerGroup) => (
+                <TableRow {...headerGroup.getHeaderGroupProps()}>
+                  {headerGroup.headers.map((column, i, arr) => (
+                    <TableCell
+                      {...column.getHeaderProps(column.getSortByToggleProps())}
+                      type="th"
+                      ref={
+                        i === 0
+                          ? firstCellRef
+                          : i === arr.length - 1
+                          ? lastCellRef
+                          : null
+                      }
+                    >
+                      <Box direction="row" justify="between">
+                        <Text weight="bold">{column.render('Header')}</Text>
+                        {column.canSort && (
+                          <SortBy
+                            isSorted={column.isSorted}
+                            isSortedDesc={column.isSortedDesc}
+                          />
+                        )}
+                      </Box>
+                      {column.canResize && (
+                        <Resizer
+                          {...column.getResizerProps()}
+                          isResizing={column.isResizing}
                         />
                       )}
-                    </Box>
-                    {column.canResize && (
-                      <Resizer
-                        // eslint-disable-next-line react/jsx-props-no-spreading
-                        {...column.getResizerProps()}
-                        isResizing={column.isResizing}
-                        onClick={(e) => e.stopPropagation()}
-                      />
-                    )}
-                    <SortByBorder isSorted={column.isSorted} />
-                  </TableCell>
-                ))}
-              </TableRow>
-            ))}
-          </TableHeader>
-          <TableBody
-            // eslint-disable-next-line react/jsx-props-no-spreading
-            {...getTableBodyProps()}
-            className="body"
-            height={tableExpanded ? tableExpandedHeight : tableDefaultHeight}
-          >
-            {rows.map((row) => {
-              prepareRow(row)
-
-              return (
-                // eslint-disable-next-line react/jsx-props-no-spreading
-                <TableRow {...row.getRowProps()} direction="row">
-                  {row.cells.map((cell) => (
-                    // eslint-disable-next-line react/jsx-props-no-spreading
-                    <TableCell {...cell.getCellProps()}>
-                      <Text
-                        style={{
-                          whiteSpace:
-                            cellWidths &&
-                            cellWidths.columnWidths?.cell?.column?.id < 100
-                              ? 'nowrap'
-                              : 'normal'
-                        }}
-                      >
-                        {cell.render('Cell')}
-                      </Text>
                     </TableCell>
                   ))}
                 </TableRow>
-              )
-            })}
-          </TableBody>
-        </Table>
-      </Box>
+              ))}
+            </TableHeader>
+            <TableBody {...getTableBodyProps()} className="body">
+              {rows.map((row) => {
+                prepareRow(row)
+                return (
+                  <TableRow {...row.getRowProps()}>
+                    {row.cells.map((cell) => (
+                      <TableCell {...cell.getCellProps()}>
+                        <Text
+                          style={{
+                            whiteSpace:
+                              cellWidths &&
+                              cellWidths.columnWidths?.cell?.column?.id < 100
+                                ? 'nowrap'
+                                : 'normal'
+                          }}
+                        >
+                          {cell.render('Cell')}
+                        </Text>
+                      </TableCell>
+                    ))}
+                  </TableRow>
+                )
+              })}
+            </TableBody>
+          </Table>
+        </BoxBlock>
+      </BoxBlock>
     </Box>
   )
 }
