+++ src/components/shared/DataTable/utils/GlobalFilter.js
@@ -1,28 +1,10 @@
-import { useState, memo } from 'react'
+import { memo } from 'react'
 import { useResponsive } from 'hooks/useResponsive'
-import { useAsyncDebounce } from 'react-table'
-import { Box, FormField, Text } from 'grommet'
-import { Icon } from 'components/shared/Icon'
-import { TextInput } from 'components/shared/TextInput'
-import { SrOnly } from 'components/shared/SrOnly'
+import { Box } from 'grommet'
+import { FilterTextInput } from 'components/shared/FilterTextInput'
 
 export const GlobalFilter = ({ globalFilter, setGlobalFilter }) => {
   const { setResponsive } = useResponsive()
-  const [userInput, setUserInput] = useState(globalFilter)
-
-  const debounceInput = useAsyncDebounce((value) => {
-    setGlobalFilter(value || '')
-  }, 500)
-
-  const handleChange = (value) => {
-    setUserInput(value)
-    debounceInput(value)
-  }
-
-  const clearInput = () => {
-    setUserInput('')
-    setGlobalFilter('')
-  }
 
   return (
     <Box
@@ -30,44 +12,11 @@ export const GlobalFilter = ({ globalFilter, setGlobalFilter }) => {
       justify="start"
       align={setResponsive('start', 'center')}
     >
-      <Text
-        margin={{ right: 'small', bottom: setResponsive('xsmall', 'none') }}
-      >
-        Filter
-      </Text>
-      <FormField
-        htmlFor="global-filter"
-        a11yTitle="Global Filter"
-        role="search"
-        width={setResponsive('100%', 'auto')}
-        style={{ position: 'relative' }}
-      >
-        <TextInput
-          id="global-filter"
-          type="text"
-          value={userInput || ''}
-          placeholder="Filter table"
-          style={{ paddingRight: '28px' }}
-          onChange={(e) => {
-            handleChange(e.target.value)
-          }}
-        />
-        {userInput && (
-          <Box
-            role="button"
-            style={{
-              boxShadow: 'none',
-              position: 'absolute',
-              right: '8px',
-              top: '8px'
-            }}
-            onClick={clearInput}
-          >
-            <Icon name="Close" size="16px" />
-            <SrOnly>Clear text</SrOnly>
-          </Box>
-        )}
-      </FormField>
+      <FilterTextInput
+        filter={globalFilter}
+        setFilter={setGlobalFilter}
+        placeholder="Filter samples"
+      />
     </Box>
   )
 }
